fix: rebase changes for flaky test fix
